+++ app/components/npq_separation/sub_navigation_component.html.erb
